chapter-1/domain-names/requirements.txt
python-whois
requests
dnspython